requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.10
django-allauth==0.32.0
django-crispy-forms==1.6.1
idna==2.5
oauthlib==2.0.2
python-openid==2.2.5
requests==2.18.2
requests-oauthlib==0.8.0
urllib3==1.22
